+++ TGS2024/GameClearScene.cpp
@@ -226,11 +226,11 @@ void GameClearScene::Draw() const
 	case 0:
 		DrawGraph(0, 0, ending_img[endimg_num], TRUE);
 
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha2);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha2);
 		DrawGraph(0, 0, ending_img[endimg_num+1], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha);
 		DrawGraph(0, 0, ending_img[7], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 		break;
@@ -238,7 +238,7 @@ void GameClearScene::Draw() const
 	case 2:
 		DrawGraph(0, 0, ending_img[endimg_num], TRUE);
 
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha2);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha2);
 		DrawGraph(0, 0, ending_img[endimg_num + 1], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 
@@ -247,7 +247,7 @@ void GameClearScene::Draw() const
 	case 4:
 	case 5:
 		DrawGraph(0, 0, ending_img[endimg_num], TRUE);
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha2);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha2);
 		DrawGraph(0, 0, ending_img[endimg_num + 1], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 		break;
@@ -255,11 +255,11 @@ void GameClearScene::Draw() const
 	case 7:
 		DrawGraph(0, 0, ending_img[endimg_num], TRUE);
 
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha);
 		DrawGraph(0, 0, ending_img[7], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, alpha2);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, (int)alpha2);
 		DrawGraph(0, 0, ending_img[8], TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 		break;
